+++ src/gui.js
@@ -91,7 +91,7 @@ modules.gui = '2025-September-22';
 
 // Declarations
 
-var SnapVersion = '11.0.4';
+var JASMVersion = '11.1-dev';
 
 var IDE_Morph;
 var ProjectDialogMorph;
@@ -1038,62 +1038,31 @@ IDE_Morph.prototype.createLogo = function () {
     // otherwise would be compromised by annoying browser security.
 
     // this.logo.texture = this.logoURL; // original code, commented out
-    this.logo.texture = "data:image/png;base64," +
-        "iVBORw0KGgoAAAANSUhEUgAAACwAAAAYCAYAAACBbx+6AAAKiklEQVRYR5VXe3BU5RX/" +
-        "ne+7924SwiOEJJvwUCAgCZFBEtRatIlVlATLIwlFsCgdeYWICu1MfbKUabVVtBoDQlUc" +
-        "FCubEIpAAEUTrGhFGIXAAjZCFdhNQiTkQbK7997vdO7SREAo9P5zZ77HOb9zzu87D8JV" +
-        "fOyBwGIwEdg5XrcmKRExcoSCNQKgWwXRTYKQDAKUQi1DbASrjzgsdqdM8zc6d6o80LIB" +
-        "RR6oq1B52SN0pcteL+SUKbCdcw3lCUMsof2amAs0iVRNEoIhZYKoCcTtYBARxUUZ1IMZ" +
-        "CIZxWDG9oVSv1/tP8Z12ZHAVNMqBdSW9l9uPAGYGoQwicqjQUQsmZ9kLSf8FGyhzzyCB" +
-        "P8X1kO7TLaoREJuIxCeSzKNhWzRbKhgyRCwJZfcA2UOY+E4QTewZK2Ob2tQhIl6cPLmu" +
-        "LKLPC+n8O2X/P+CJAXLAXXzpfLD+sqRHesaKF5vbHZtil4bCA98YeO+2f19J0Yl3+wzV" +
-        "DH0GMz8cE0WxHSH8DZrxhPsX3x7rBO5YUFgI1Um3y8r0sCg8WOZgBQ54YPTJGNCPgehw" +
-        "qNl/zfTmJoe3Dt9OeN15LgObTUs/JNB9prvA9/mljNvblCkyh+7l6p3AxVxt2JiQalty" +
-        "IYB5AL5n5qWh1vqVA2cieCWjz+07AXd8C+eZAP71SY8Q6JlzfuajDPFMSkHg7brtSd1w" +
-        "Vr2hVIymxX97f2IO2nCPP2be0EDaWZuMVttoP2tGBd5/dfCpToHnKMZUvWSJzP5ZNSin" +
-        "uouv9RXX/MRW9lMgHkekaqCsVZDmZnfD4JMI7LXPPUgHXATaBVEvLDrg7tBgRDbrK9wz" +
-        "GHwnM0Xrmsg3bT4eC5XV2FzfYnS/fkzK9zU7aQ7MXxbvnxkk8UhYUTcGTGJyMsM/Okw5" +
-        "s3rVdY2Zs/foe1MyIw8UHjA8oCosEUA1cjw/AA94M/KUMOcQBW8gsptYuXYpa8Cr/aZW" +
-        "7Sss9Mrhw33swWJkV1eL6uoc6wFPVVRDo3stmDN/xOFAed95EHYps7o/Jb/hrc6QTXt0" +
-        "/4QzYa1Egd7TyCq3WEgBGkggMyGhbt2bnpyrDO8PJDizAYPbbS21Tw+rXk+BjzIQvhRF" +
-        "8ub6MlhiF4h6dKU1J1M4xD+xvnc/CaMKpN5LntywqHM9d77vrwCNrCxNG32x0Oxs1lzp" +
-        "vmtdQVnfe0DArGvsczNskUAaareWDP/SOT+2qKa/DkrtLu14k8HrW+JrsKbf1xFZN3ES" +
-        "khrbJ7tPxYYMMRpsxQi4ajaVDjnobI8vrslWLLc6186lNYBqX041hiyoDR339ovWNGs7" +
-        "GA3J+XUFneDGFft+T4zfCsYDm5enrzsfdF7R12lM1jsAfcPgNmJkMqE3AfEMWqYTlVpK" +
-        "vcDAbSCcEUCcIO6jSyzWSW04a8rXmGAw4yQYg5nQkxi9GHhu6/L0pbnzfbcxoZIUFXd5" +
-        "2KlEOR5Yfm/cACFduxnCl5zvv70TWN68/YNYauVSi77BNjs2CmDVQKF/WFIyJPTzh48m" +
-        "GVbwCwK6E+MJJtpBLKUi+1kC3wNShbaF40KDrkM7FrQ0S5PmsyCMd5xAzHMVYRgzzbCV" +
-        "/jkb4Z66En/WpGuisjryFIkGsFqrWN0XAXx+NQuUpyyJ70VPnz5jfapc7RNS7mltXLly" +
-        "tj5nzipzbPG+gTrrTzIwQ2guTZmhHUoXxdteGnYkd/6hfUR8cMsr6dM6jcwt+nokkbkL" +
-        "JBdseWXY6+dH5a6iw3dLUiuYsQJEPwXQurU07b7OM3c9ery3DLceAdHHgvl1xVQYIvzG" +
-        "AUzshXCqTsP65NtsxioQWgAVw2w/kFLQuGfPykw9a84eqzPV3D2vZgQJ7UEp9YfYDtXa" +
-        "mhwvLHs5QTRvKU2b3AW4+ND1YOwQQi3cXDJ8be78QwsZGCXAUgFDCdRPET8uGGMBiqlM" +
-        "WDcBHo9yMkVZ2RQ7d75vEzMGMMmFUqqO0b2H/dMBGym/zBB1Fe6PwBAgvAxgBYMWpuQH" +
-        "3nLq/5KdrA42f+Y69WXIdFKNA2pcsW+iYLzDjBIQZwHUWlmaNqnTsNzimiywtoFhL2PI" +
-        "YQTOZfDbAH1B4CwCTSfiJxXTHQTun5gQk/emZ2Aw3XPA8HkywuOKfZXElFJZmjYykik9" +
-        "LLrSWl1F0iyXIVaFgmqa5rI+NsO680LXJufXzedIo3ZhIv/Bi75qAvwMpEChrnJ52r1d" +
-        "kSg6MlqStYZBxwFKZ4XpW1ek7XTuTiiq6W+SfA/Ez4FxB0EkbylNG3fem4ljoR1hoFLY" +
-        "eJ50Kdtq/AcjHG7cFN/XDOu7AWpOzg+kH/DCiJdJXzFLocX7s5wK9+CivZnfne3WM0rD" +
-        "4ZYwhWO7dbjskD6VSPwOij1MmE2E+srS9LFdmWXu4dtJU2VgOgxgqFDqKc0V827YDCaC" +
-        "uIgYs1hxMQTdAubbFctJ21YM2z95ti85aGA5gFGsuISIHgNwshurWyKAAxXJy7q5sLA1" +
-        "qGb1za9/zVnzlyeu6h7TbdbZjmNT3flYN3XBvj+22noRA8cY6CBCFJgSFdQaM6ReMlyi" +
-        "nEDHKkvTZ3R5f77vTmIuZYlXSNEoEPKZcRiMehAsJ4URsEIJSiPmOQT+EKAWJhoEcIKm" +
-        "xFxbKottVICwrrI0fTY5Pa5N8iunh2i3w2MGT2lqdhTWlSWNj4kxNp0Nth8Qoe/vSCph" +
-        "c2rWgYk2EE8gYZNqs1l88feSjN0RPj908AZlo3X78uG1nYBnPHYoHh0dQweh+ZCzdgjx" +
-        "eU5B0Q0+2MduOtAsY+Paw3qo1daeAXFSFJnLJIm+LIi6a+Hq1ctG+bwvfBq97pueg4TR" +
-        "42jZi/07KFDh9ib20gpPnbH/4J4ceHLPSuhZc2AeW31tVFT34Fp3ojE50Gi9n5zqn0oj" +
-        "0HSp0nmpNY/HIzwez1VNF+OLD35gM4W3lqbn/W/5TBRYn7iISPaxFXn7Fvi/9Hgg0tNB" +
-        "zpRR571mIMtgSbcokXe2PcavKLaCYR4DFBT1qvWfnFZ984IFLU4rugRVoroaqKrKsZ0e" +
-        "0OmxT3qzrlOC7pZojmbWmcggWylACNh2nBYb9VG4LTy9ZuqOJY/31my9dMziF3vGvDug" +
-        "pSPb0GWzBdkEwWSdbs/aOPxXZZHIXTAidTbzzj9Srwns35QSgzDfJdjKBon+DM1m5gwi" +
-        "dAjhL0yahG/+VZnqSt1dazoC9yZDZs6G5dwNbEhcBIXHAdpFZCu2NQ0kmahdWZyoubQj" +
-        "aLMmbc/Z9pdR6a4Qv5bzYK2ufTwmZGUoTXxnsooxGByWetPTSRPC+yN9zeVC4OBd4gF5" +
-        "zhsanUY/w4PwiQ19R0plvQWmpckFdd7Lyagrd29i4Nvkgrpix/DTHaboHa1HaCKMDFLh" +
-        "9/lIo0c9/dmUOKkpXj36+TOuPm+KU8ZYSggfYGHYpMKSP+nwhzrnSnLCWZYOutyYEpm/" +
-        "fOCLp9268uQXQOpGZnKKTBtLinaYAgJJojZWfCsDBSTlFPfEEzVXy/3/5UCHZlecmh0B" +
-        "jrfLvBAJPlC/G1PlkNza0OkP4noGW4zVhkaTTAsWsTNnkDP02XSu82oTTPOSCgJvOw85" +
-        "0xE09MezY9mpQp7i87IHwOJ0IiRcSNOIAdkRmZEJ5D9/VBCtnsd7nAAAAABJRU5ErkJg" +
-        "gg==";
+    this.logo.texture = "data:image/png;base64," + 
+        "iVBORw0KGgoAAAANSUhEUgAAACwAAAAYCAYAAACBbx+6AAAAAXNSR0IB2cksfwAAAAl" + 
+        "wSFlzAAALEwAACxMBAJqcGAAABERJREFUeJzF139oVlUcx/G5X2maQ53b1OzXcGpa6T" + 
+        "Lbytx0ouIsRPFHgkTpbpYFKSJFQVrNkpYVWUJoSQjNzaXmH4mYKaKSaSP7YaXoyLhSa" + 
+        "Zql81db7y/3c+lyfJ5HHzf0gRfjnnvPOZ977rnnnqWkJPnz63pch5FYiE+xD3/ijPyG" + 
+        "r7AIw+36ZPtolR8dX49RqMUvCncUP6JeIXfjexzBeTSgGoPQ9mqGzcQCHEczDmExpqI" + 
+        "QNyMX3dAP47BcN2bX78dTV2W06SQVa9Xxz5iANpdZN3wqNvLn8B3GtHi0aSAPAzRaPc" + 
+        "NA9hfTcBI1uNduIMm2TW88h5806vZ0+lzujcdq9BU95sP4DP1V3hE7sRQ3XFHj//dhT" + 
+        "6oI6zQAuzQY7a+ksekKXa03/gDSkYWClgSN09+tmI+9OIXNKE2mgXSNQLaC/+sG9bzm" + 
+        "tuiENpGyDHRGO7dNK0N3SbOyiopmK++IDvUrRtjy2BfD8EBDTUFueC7sg783YRK6Jgp" + 
+        "/PxpR7gR4BB8hT8epGIuVmB2G0rn2eBXr5WGV9cYyLMVgp/0inVtk16lsOpoxJlHgfM" + 
+        "3lZ5wGP1DlfjrOxw78jQYMV7mN0Fs4ggV4ElXogWlq4yT2Iz3S/mb8LpNUFl4/IlHge" + 
+        "3BMX6rMBIEfwiksVieVKu+GdahHoT1ee6SaPmGAjbhg51XnRt34O07gybq+JF7YFL0M" + 
+        "tuauROcEgavUyS3Yi52REX4b51GLcs9rCudkGNimy3HMU/k8HMVoJ/B4XV8cL/BE/IB" + 
+        "t+NrWz1iB0UUhP9Y5mwJNuE3HNl/fha86dXqBwsD2PmzCt/ZOYJ/mdakTuFzXF8YKa5" + 
+        "/frfgcz+OEvb1xAhfjhB77+5rLdm4GUnV9FsqwFmcxJRLYHvVMnMMcTa0JMQIPwSEUu" + 
+        "GHNEC1ns/1gN2bTYkZFRVMYuFqd9cV8/IWtGqltOGYjjp42mpEbDeeh5wTOxmlNjUYd" + 
+        "u4Ft9B+3J+YG7o9vsB7dtQc4izWznmi0peuAXpLDyFHYFayr4dzMxMv4B09jD3bhC/z" + 
+        "hBStAfjSw6q3W8Rs6dgN7Ov9gNKx9NF7CBc3hNJXbV+hgTVVlCRWWe8EaOVIj8Z7nLD" + 
+        "VesCLYPBzmBS/Ph1iFN/VU0jAAS+yv6ozS8R06tpuqxJ06Hqx+e0UD3+UHe1hbyjpEy" + 
+        "pfpJp69aMJfy592YbbZrnLK7/aD/yZsLj96rfJd9CPMJnzpfoo5zsAcbU5+xUxtE3O1" + 
+        "57Cd3KXatv1CjurEkiPZ0jVSFtUu2miBH+yFs2J0aKFsH3vQDzbge/CJdnUvRr+EcQL" + 
+        "fh9WqE6qLsHO1as+sUpmrLNmnkKIRsQ2+/WP5GLbg9kvUs5u1fW+Z6tnOrBQlGGrrvB" + 
+        "9stoqkWOVDI+dNXlKBYwSxffIGzPWDD066I0PLo23WF7aos9b6EeQFTZXX8ZrDNk5rs" + 
+        "B0DW7Pf/wDSaN5sGMXxKwAAAABJRU5ErkJggg==";
 
     this.logo.render = function (ctx) {
         var gradient = ctx.createLinearGradient(
@@ -1604,7 +1573,7 @@ IDE_Morph.prototype.createControlBar = function () {
         name = (myself.getProjectName() || localize('untitled'));
         if (!myself.config.preserveTitle) {
             document.title = "Snap! " +
-                (myself.getProjectName() ? name : SnapVersion);
+                (myself.getProjectName() ? name : JASMVersion);
         }
         txt = new StringMorph(
             prefix + name +  scene + suffix,
@@ -5616,7 +5585,8 @@ IDE_Morph.prototype.aboutSnap = function () {
         module, btn1, btn2, btn3, btn4, licenseBtn, translatorsBtn,
         world = this.world();
 
-    aboutTxt = 'Snap! ' + SnapVersion + '\nBuild Your Own Blocks\n\n'
+    aboutTxt = 'JASM! ' + JASMVersion + '\nMod of Snap!\n\n'
+        + 'by codingisfun2831t\n\n'
         + 'Copyright \u24B8 2008-2025 Jens M\u00F6nig and '
         + 'Brian Harvey\n'
         + 'jens@moenig.org, bh@cs.berkeley.edu\n\n'
@@ -9036,7 +9006,7 @@ IDE_Morph.prototype.isIE = function () {
 // IDE_Morph warn about saving project in the dev version
 
 IDE_Morph.prototype.warnAboutDev = function () {
-    if (!SnapVersion.includes('-dev') || this.config.noDevWarning) {
+    if (!JASMVersion.includes('-dev') || this.config.noDevWarning) {
         return;
     }
     this.inform(